+++ docs/changelog.rst
@@ -17,6 +17,7 @@ New features
 
   * New attribute :attr:`aspecd.tasks.PlotTask.target` allows adding a plot to an already existing plot.
   * :meth:`aspecd.tasks.Task.to_dict` adds (implicit) parameters of underlying task object
+  * Classes from the ASpecD framework can be used without prefixing them with "aspecd" in recipes with "default_package" set to a package based on the ASpecD framework.
 
 
 Version 0.3.1